third_party/blink/web_tests/platform/mac-mac14-arm64/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/conv_transpose2d.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] [required] convTranspose2d same output size different padding (padding=2, outputPadding=2))
  assert_less_than_equal: assert_array_approx_equals_ulp: test convTranspose2d float32 actual 1 should be close enough to expected 2 by ULP distance: expected a number less than or equal to 18n but got 8388608n
Harness: the test ran to completion.
